requirements.txt
discord.py[speed]
requests
lxml
cfscrape
beautifulsoup4